@@ -20,7 +20,10 @@
     ],
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=5.3.0"
+        "php": "^7.1||^8.0"
+    },
+    "suggest": {
+        "paragonie\/sodium_compat": "Support EdDSA (Ed25519) signatures when libsodium is not present"
     },
     "autoload": {
         "psr-4": {
@@ -28,6 +31,6 @@
         }
     },
     "require-dev": {
-        "phpunit\/phpunit": ">=4.8 <=9"
+        "phpunit\/phpunit": "^7.5||9.5"
     }
 }
