@@ -10,23 +10,24 @@
     "homepage": "http:\/\/github.com\/google\/google-auth-library-php",
     "license": "Apache-2.0",
     "support": {
-        "docs": "https:\/\/googleapis.github.io\/google-auth-library-php\/master\/"
+        "docs": "https:\/\/googleapis.github.io\/google-auth-library-php\/main\/"
     },
     "require": {
-        "php": ">=5.4",
-        "firebase\/php-jwt": "~2.0|~3.0|~4.0|~5.0",
-        "guzzlehttp\/guzzle": "^5.3.1|^6.2.1|^7.0",
-        "guzzlehttp\/psr7": "^1.2",
+        "php": "^7.1||^8.0",
+        "firebase\/php-jwt": "^5.5||^6.0",
+        "guzzlehttp\/guzzle": "^6.2.1|^7.0",
+        "guzzlehttp\/psr7": "^1.7|^2.0",
         "psr\/http-message": "^1.0",
-        "psr\/cache": "^1.0"
+        "psr\/cache": "^1.0|^2.0|^3.0"
     },
     "require-dev": {
         "guzzlehttp\/promises": "0.1.1|^1.3",
         "squizlabs\/php_codesniffer": "^3.5",
-        "phpunit\/phpunit": "^4.8.36|^5.7",
+        "phpunit\/phpunit": "^7.5||^8.5",
+        "phpspec\/prophecy-phpunit": "^1.1",
         "sebastian\/comparator": ">=1.2.3",
-        "phpseclib\/phpseclib": "^2",
-        "kelvinmo\/simplejwt": "^0.2.5"
+        "phpseclib\/phpseclib": "^2.0.31",
+        "kelvinmo\/simplejwt": "^0.2.5|^0.5.1"
     },
     "suggest": {
         "phpseclib\/phpseclib": "May be used in place of OpenSSL for signing strings or for token management. Please require version ^2."
