@@ -1,6 +1,5 @@
 {
     "name": "guzzlehttp\/psr7",
-    "type": "library",
     "description": "PSR-7 message implementation that also provides common utility methods",
     "keywords": [
         "request",
@@ -15,26 +14,55 @@
     "license": "MIT",
     "authors": [
         {
+            "name": "Graham Campbell",
+            "email": "hello@gjcampbell.co.uk",
+            "homepage": "https:\/\/github.com\/GrahamCampbell"
+        },
+        {
             "name": "Michael Dowling",
             "email": "mtdowling@gmail.com",
             "homepage": "https:\/\/github.com\/mtdowling"
         },
         {
+            "name": "George Mponos",
+            "email": "gmponos@gmail.com",
+            "homepage": "https:\/\/github.com\/gmponos"
+        },
+        {
+            "name": "Tobias Nyholm",
+            "email": "tobias.nyholm@gmail.com",
+            "homepage": "https:\/\/github.com\/Nyholm"
+        },
+        {
+            "name": "M\u00e1rk S\u00e1gi-Kaz\u00e1r",
+            "email": "mark.sagikazar@gmail.com",
+            "homepage": "https:\/\/github.com\/sagikazarmark"
+        },
+        {
             "name": "Tobias Schultze",
+            "email": "webmaster@tubo-world.de",
             "homepage": "https:\/\/github.com\/Tobion"
+        },
+        {
+            "name": "M\u00e1rk S\u00e1gi-Kaz\u00e1r",
+            "email": "mark.sagikazar@gmail.com",
+            "homepage": "https:\/\/sagikazarmark.hu"
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "psr\/http-message": "~1.0",
-        "ralouphie\/getallheaders": "^2.0.5 || ^3.0.0"
-    },
-    "require-dev": {
-        "phpunit\/phpunit": "~4.8.36 || ^5.7.27 || ^6.5.14 || ^7.5.20 || ^8.5.8 || ^9.3.10",
-        "ext-zlib": "*"
+        "php": "^7.2.5 || ^8.0",
+        "psr\/http-factory": "^1.0",
+        "psr\/http-message": "^1.0",
+        "ralouphie\/getallheaders": "^3.0"
     },
     "provide": {
+        "psr\/http-factory-implementation": "1.0",
         "psr\/http-message-implementation": "1.0"
+    },
+    "require-dev": {
+        "bamarni\/composer-bin-plugin": "^1.4.1",
+        "http-interop\/http-factory-tests": "^0.9",
+        "phpunit\/phpunit": "^8.5.8 || ^9.3.10"
     },
     "suggest": {
         "laminas\/laminas-httphandlerrunner": "Emit PSR-7 responses"
@@ -42,10 +70,7 @@
     "autoload": {
         "psr-4": {
             "DeliciousBrains\\WP_Offload_Media\\Gcp\\GuzzleHttp\\Psr7\\": "src\/"
-        },
-        "files": [
-            "src\/functions_include.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
@@ -54,7 +79,14 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.7-dev"
+            "dev-master": "2.2-dev"
         }
+    },
+    "config": {
+        "allow-plugins": {
+            "bamarni\/composer-bin-plugin": true
+        },
+        "preferred-install": "dist",
+        "sort-packages": true
     }
 }
