@@ -87,7 +87,7 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 
 		/**
 		 * @param string      $plugin_name
-		 * @param  string     $plugin_slug
+		 * @param string      $plugin_slug
 		 * @param string      $plugin_file_path
 		 * @param string|null $parent_plugin_name
 		 * @param string|null $parent_plugin_slug
@@ -132,7 +132,7 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 		 * @return bool
 		 */
 		function is_plugin_active( $plugin_base ) {
-			include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
+			include_once ABSPATH . 'wp-admin/includes/plugin.php';
 
 			return is_plugin_active( $plugin_base );
 		}
@@ -314,8 +314,8 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 
 			$plugin_basename = $this->get_plugin_basename();
 			$deactivate_url  = $this->get_plugin_action_url( 'deactivate', $plugin_basename );
-			$deactivate_link = sprintf( '<a style="text-decoration:none;" href="%s">%s</a>', $deactivate_url, __( 'deactivate', 'amazon-s3-and-cloudfront' ) );
-			$hide_notice_msg = '<br><em>' . sprintf( __( 'You can %s the %s plugin to get rid of this notice.', 'amazon-s3-and-cloudfront' ), $deactivate_link, $this->plugin_name ) . '</em>';
+			$deactivate_link = sprintf( '<a style="text-decoration:none;" href="%s">%s</a>', $deactivate_url, __( 'deactivate' ) );
+			$hide_notice_msg = '<br><em>' . sprintf( __( 'You can %s the %s plugin to get rid of this notice.' ), $deactivate_link, $this->plugin_name ) . '</em>';
 
 			// Check basic requirements for AWS SDK.
 			$sdk_errors = $this->get_sdk_requirements_errors();
@@ -334,22 +334,22 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 			$parent_plugin_link_html = sprintf( '<a style="text-decoration:none;" href="%s">%s</a>', $this->get_parent_plugin_url(), $this->get_parent_plugin_name() );
 
 			if ( ! $this->is_parent_plugin_enabled() ) {
-				$msg = sprintf( __( '%s has been disabled as it requires the %s plugin.', 'amazon-s3-and-cloudfront' ), $this->plugin_name, $parent_plugin_link_html );
+				$msg = sprintf( __( '%s has been disabled as it requires the %s plugin.' ), $this->plugin_name, $parent_plugin_link_html );
 
 				if ( file_exists( WP_PLUGIN_DIR . '/' . $parent_basename ) ) {
 					if ( isset( $GLOBALS['aws_meta'][ $this->parent_plugin_slug ]['compatible'] ) && ! $GLOBALS['aws_meta'][ $this->parent_plugin_slug ]['compatible'] ) {
-						$msg = rtrim( $msg, '.' ) . ', ' . __( 'which is currently disabled.', 'amazon-s3-and-cloudfront' );
+						$msg = rtrim( $msg, '.' ) . ', ' . __( 'which is currently disabled.' );
 					} else {
-						$msg          .= ' ' . __( 'It appears to be installed already.', 'amazon-s3-and-cloudfront' );
+						$msg          .= ' ' . __( 'It appears to be installed already.' );
 						$activate_url = $this->get_plugin_action_url( 'activate', $parent_basename );
-						$msg          .= ' <a id="' . $this->plugin_slug . '-activate-parent" style="font-weight:bold;text-decoration:none;" href="' . $activate_url . '">' . _x( 'Activate it now.', 'Activate plugin', 'amazon-s3-and-cloudfront' ) . '</a>';
+						$msg          .= ' <a id="' . $this->plugin_slug . '-activate-parent" style="font-weight:bold;text-decoration:none;" href="' . $activate_url . '">' . _x( 'Activate it now.', 'Activate plugin' ) . '</a>';
 					}
 				} else {
 					$install_url = 'https://deliciousbrains.com/my-account/';
 					if ( is_null( $this->parent_plugin_url ) ) {
 						$install_url = $this->get_plugin_action_url( 'install', $this->parent_plugin_slug );
 					}
-					$msg .= ' ' . sprintf( __( '<a href="%s">Install</a> and activate it.', 'amazon-s3-and-cloudfront' ), $install_url );
+					$msg .= ' ' . sprintf( __( '<a href="%s">Install</a> and activate it.' ), $install_url );
 				}
 
 				$msg .= $hide_notice_msg;
@@ -360,17 +360,17 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 			$current_parent_plugin_version = isset( $GLOBALS['aws_meta'][ $this->parent_plugin_slug ]['version'] ) ? $GLOBALS['aws_meta'][ $this->parent_plugin_slug ]['version'] : 0;
 
 			if ( ! version_compare( $current_parent_plugin_version, $this->parent_plugin_required_version, '>=' ) ) {
-				$msg = sprintf( __( '%s has been disabled as it requires version %s or later of the %s plugin.', 'amazon-s3-and-cloudfront' ), $this->plugin_name, $this->parent_plugin_required_version, $parent_plugin_link_html );
+				$msg = sprintf( __( '%s has been disabled as it requires version %s or later of the %s plugin.' ), $this->plugin_name, $this->parent_plugin_required_version, $parent_plugin_link_html );
 
 				if ( $current_parent_plugin_version ) {
-					$msg .= ' ' . sprintf( __( 'You currently have version %s installed.', 'amazon-s3-and-cloudfront' ), $current_parent_plugin_version );
+					$msg .= ' ' . sprintf( __( 'You currently have version %s installed.' ), $current_parent_plugin_version );
 				}
 
 				global $as3cfpro;
 				if ( ! empty( $as3cfpro ) && $as3cfpro->get_plugin_slug( true ) === $this->parent_plugin_slug ) {
 					// Don't show update link for addons of a licensed plugin where the license is invalid
 					if ( ! $as3cfpro->is_valid_licence( false, true ) ) {
-						$msg .= ' ' . sprintf( __( 'A valid license for %s is required to update.', 'amazon-s3-and-cloudfront' ), $this->get_parent_plugin_name() );
+						$msg .= ' ' . sprintf( __( 'A valid license for %s is required to update.' ), $this->get_parent_plugin_name() );
 						$msg .= $hide_notice_msg;
 
 						return $this->set_error_msg( $msg );
@@ -379,7 +379,7 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 
 				$update_url = $this->get_plugin_action_url( 'upgrade', $parent_basename );
 
-				$msg .= ' <a style="font-weight:bold;text-decoration:none;white-space:nowrap;" href="' . $update_url . '">' . __( 'Update to the latest version', 'amazon-s3-and-cloudfront' ) . '</a>';
+				$msg .= ' <a style="font-weight:bold;text-decoration:none;white-space:nowrap;" href="' . $update_url . '">' . __( 'Update to the latest version' ) . '</a>';
 				$msg .= $hide_notice_msg;
 
 				return $this->set_error_msg( $msg );
@@ -390,7 +390,7 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 			}
 
 			if ( ! isset( $GLOBALS['aws_meta'][ $this->parent_plugin_slug ]['supported_addon_versions'][ $this->plugin_slug ] ) ) {
-				$msg = sprintf( __( '%1$s has been disabled because it is not a supported addon of the %2$s plugin.', 'amazon-s3-and-cloudfront' ), $this->plugin_name, $this->get_parent_plugin_name() );
+				$msg = sprintf( __( '%1$s has been disabled because it is not a supported addon of the %2$s plugin.' ), $this->plugin_name, $this->get_parent_plugin_name() );
 
 				return $this->set_error_msg( $msg );
 			}
@@ -399,16 +399,16 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 			$this_plugin_version          = $GLOBALS['aws_meta'][ $this->plugin_slug ]['version'];
 
 			if ( ! version_compare( $this_plugin_version, $this_plugin_version_required, '>=' ) ) {
-				$msg = sprintf( __( '%1$s has been disabled because it will not work with the version of the %2$s plugin installed. %1$s %3$s or later is required.', 'amazon-s3-and-cloudfront' ), $this->plugin_name, $this->get_parent_plugin_name(), $this_plugin_version_required );
+				$msg = sprintf( __( '%1$s has been disabled because it will not work with the version of the %2$s plugin installed. %1$s %3$s or later is required.' ), $this->plugin_name, $this->get_parent_plugin_name(), $this_plugin_version_required );
 
 				$update_url  = $this->get_plugin_action_url( 'upgrade', $plugin_basename );
-				$upgrade_msg = ' <a style="font-weight:bold;text-decoration:none;white-space:nowrap;" href="' . $update_url . '">' . sprintf( __( 'Update %s to the latest version', 'amazon-s3-and-cloudfront' ), $this->plugin_name ) . '</a>';
+				$upgrade_msg = ' <a style="font-weight:bold;text-decoration:none;white-space:nowrap;" href="' . $update_url . '">' . sprintf( __( 'Update %s to the latest version' ), $this->plugin_name ) . '</a>';
 
 				global $as3cfpro;
 				if ( ! empty( $as3cfpro ) && $as3cfpro->get_plugin_slug( true ) === $this->parent_plugin_slug ) {
 					// Don't show update link for addons of a licensed plugin where the license is invalid
 					if ( ! $as3cfpro->is_valid_licence( false, true ) ) {
-						$upgrade_msg = ' ' . sprintf( __( 'A valid license for %s is required to update.', 'amazon-s3-and-cloudfront' ), $this->get_parent_plugin_name() );
+						$upgrade_msg = ' ' . sprintf( __( 'A valid license for %s is required to update.' ), $this->get_parent_plugin_name() );
 					}
 				}
 
@@ -471,7 +471,7 @@ if ( ! class_exists( 'AS3CF_Compatibility_Check' ) ) {
 			}
 
 			if ( $this->deactivate_if_not_compatible ) {
-				$deactivated_msg = sprintf( __( 'The %s plugin has been deactivated.', 'amazon-s3-and-cloudfront' ), $this->plugin_name );
+				$deactivated_msg = sprintf( __( 'The %s plugin has been deactivated.' ), $this->plugin_name );
 
 				$error_msg = $deactivated_msg . ' ' . $error_msg;
 				$this->render_notice( $error_msg );
