@@ -45,7 +45,6 @@ class AS3CF_S3_To_Local extends AS3CF_Filter {
 	 * @param array $instance
 	 *
 	 * @return array
-	 *
 	 */
 	public function filter_widget_save( $instance ) {
 		return $this->handle_widget( $instance );
@@ -247,7 +246,10 @@ class AS3CF_S3_To_Local extends AS3CF_Filter {
 	 * @return string
 	 */
 	protected function post_process_content( $content ) {
-		return $this->remove_aws_query_strings( $content );
+		$content = $this->remove_aws_query_strings( $content );
+		$content = AS3CF_Utils::maybe_fix_serialized_string( $content );
+
+		return $content;
 	}
 
 	/**
