@@ -150,6 +150,7 @@ class Remove_Local_Handler extends Item_Handler {
 
 			$file_to_remove['remove_result'] = array( 'status' => self::STATUS_OK );
 
+			//phpcs:ignore
 			if ( ! @unlink( $file ) ) {
 				$this->remove_blocked[] = $file;
 
@@ -158,7 +159,7 @@ class Remove_Local_Handler extends Item_Handler {
 
 				if ( ! file_exists( $file ) ) {
 					$file_to_remove['remove_result']['message'] = "Error removing local file. Couldn't find the file at $file";
-				} else if ( ! is_writable( $file ) ) {
+				} elseif ( ! is_writable( $file ) ) {
 					$file_to_remove['remove_result']['message'] = "Error removing local file. Ownership or permissions are mis-configured for $file";
 				}
 			}
