@@ -2,7 +2,8 @@
 
 namespace DeliciousBrains\WP_Offload_Media\Upgrades;
 
-use DeliciousBrains\WP_Offload_Media\Pro\Sidebar_Presenter;
+use DeliciousBrains\WP_Offload_Media\Pro\Tool;
+use DeliciousBrains\WP_Offload_Media\Pro\Tools_Manager;
 
 /**
  * Upgrade_Tools_Errors Class
@@ -34,7 +35,7 @@ class Upgrade_Tools_Errors extends Upgrade {
 	 * @return string
 	 */
 	protected function get_running_update_text() {
-		return __( 'and reformatting internal data about previous errors from tools .', 'amazon-s3-and-cloudfront' );
+		return __( 'and reformatting internal data about previous errors from tools.', 'amazon-s3-and-cloudfront' );
 	}
 
 	/**
@@ -53,8 +54,8 @@ class Upgrade_Tools_Errors extends Upgrade {
 			return array();
 		}
 
-		$sidebar_presenter = Sidebar_Presenter::get_instance( $as3cf );
-		$tools             = $sidebar_presenter->get_all_tools();
+		$tools_manager = Tools_Manager::get_instance( $as3cf );
+		$tools         = $tools_manager->get_tools();
 
 		return array_keys( $tools );
 	}
@@ -75,10 +76,11 @@ class Upgrade_Tools_Errors extends Upgrade {
 			return true;
 		}
 
-		$sidebar_presenter = Sidebar_Presenter::get_instance( $as3cf );
-		$tools             = $sidebar_presenter->get_all_tools();
+		$tools_manager = Tools_Manager::get_instance( $as3cf );
+		$tools         = $tools_manager->get_tools();
 
 		if ( ! empty( $tools[ $item ] ) ) {
+			/** @var Tool $tool */
 			$tool = $tools[ $item ];
 
 			$errors     = $tool->get_errors();
